requirements.txt
yfinance
pandas
google-cloud-bigquery
pyarrow
pyyaml